requirements.txt
#external requirements
flask
flask-cors
cianparser
pandas
boto3
python-dotenv